requirements.txt
torch
torchvision
timm
numpy
Pillow
tqdm
scipy
tensorboard